requirements.txt
python-dotenv 
psycopg2